+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## vNext
+
+* Add quadkey interpolation for `{q}`
+
 ## v0.3.0 - 7/5/14
 
 * Update `tilejson` dependency to `^0.8.0'.